Fastapi/requirements.txt
fastapi~=0.135.1
motor~=3.7.1
redis~=7.4.0
python-jose~=3.5.0
passlib~=1.7.4
pydantic~=2.12.5
uvicorn~=0.42.0
starlette~=0.52.1
argon2-cffi>=21.1.0
bcrypt>=3.2.0